Move some common helpers to new files utils.c, utils.h
