+++ functions/Autgrp/perf_normal.c
@@ -3,6 +3,7 @@
 /*****	Main program for the automorphism program AUTO	*****/
 
 #include "typedef.h"
+#include "utils.h"
 #include "types.h"
 #include "matrix.h"
