+++ functions/Autgrp/pr_isom.c
@@ -9,6 +9,7 @@
 \**************************************************************************/
 
 #include "typedef.h"
+#include "utils.h"
 #include "types.h"
 #include "matrix.h"
 #include "autgrp.h"
