+++ functions/Contrib/suche_kand.c
@@ -1,4 +1,5 @@
 #include "typedef.h"
+#include "utils.h"
 #include "contrib.h"
 #include "matrix.h"
 #include "getput.h"
