+++ functions/Graph/sub-k-groups-fcts.c
@@ -1,22 +1,23 @@
 /* last change: 07.02.2001 by Oliver Heidbuechel */
 
 
-#include <ZZ.h>
-#include <typedef.h>
-#include <presentation.h>
-#include <matrix.h>
-#include <bravais.h>
-#include <base.h>
-#include <datei.h>
-#include <graph.h>
-#include <gmp.h>
-#include <zass.h>
-#include <tools.h>
-#include <longtools.h>
-#include <voronoi.h>
-#include <symm.h>
-#include <autgrp.h>
-#include <reduction.h>
+#include "ZZ.h"
+#include "typedef.h"
+#include "utils.h"
+#include "presentation.h"
+#include "matrix.h"
+#include "bravais.h"
+#include "base.h"
+#include "datei.h"
+#include "graph.h"
+#include "gmp.h"
+#include "zass.h"
+#include "tools.h"
+#include "longtools.h"
+#include "voronoi.h"
+#include "symm.h"
+#include "autgrp.h"
+#include "reduction.h"
 
 #define MYSIZE 1024
