+++ functions/Hyperbolic/hyp_isom.c
@@ -1,4 +1,5 @@
 #include "typedef.h"
+#include "utils.h"
 #include "autgrp.h"
 #include "matrix.h"
