+++ functions/Hyperbolic/hyp_stabilizer.c
@@ -1,4 +1,5 @@
 #include "typedef.h"
+#include "utils.h"
 #include "autgrp.h"
 #include "matrix.h"
 #include "datei.h"
