+++ functions/Matrix/p_lse_solve.c
@@ -1,4 +1,5 @@
 #include "typedef.h"
+#include "utils.h"
 #include "matrix.h"
 #include "tools.h"
