+++ functions/Matrix/p_mat_det.c
@@ -1,4 +1,5 @@
 #include "typedef.h"
+#include "utils.h"
 #include "tools.h"
 
 /**************************************************************************\
