+++ functions/Name/compute_q_matrix.c
@@ -1,4 +1,5 @@
 #include "typedef.h"
+#include "utils.h"
 #include "matrix.h"
 #include "idem.h"
 #include "bravais.h"
