+++ functions/Orbit/orbit_alg.c
@@ -1,4 +1,5 @@
 #include "typedef.h"
+#include "utils.h"
 #include "longtools.h"
 #include "matrix.h"
 #include "sort.h"
