+++ functions/Reduction/mink_red.c
@@ -1,4 +1,5 @@
-#include"typedef.h"
+#include "typedef.h"
+#include "utils.h"
 #include"matrix.h"
 #include"symm.h"
 #include"longtools.h"
