+++ functions/Reduction/short_reduce.c
@@ -1,4 +1,5 @@
-#include"typedef.h"
+#include "typedef.h"
+#include "utils.h"
 #include"matrix.h"
 #include"symm.h"
 #include"reduction.h"
