+++ functions/Symm/rest_short.c
@@ -1,4 +1,5 @@
 #include "typedef.h"
+#include "utils.h"
 #include "matrix.h"
 
 /**************************************************************************\
