+++ functions/Voronoi/all_vor_neighbours.c
@@ -1,4 +1,5 @@
 #include "typedef.h"
+#include "utils.h"
 #include "matrix.h"
 #include "voronoi.h"
 #include "polyeder.h"
