+++ functions/Voronoi/normalizer.c
@@ -1,4 +1,5 @@
 #include "typedef.h"
+#include "utils.h"
 #include "voronoi.h"
 #include "longtools.h"
 #include "symm.h"
