+++ functions/Voronoi/vor_vertices.c
@@ -1,4 +1,5 @@
 #include "typedef.h"
+#include "utils.h"
 #include "matrix.h"
 #include "orbit.h"
 #include "symm.h"
