+++ include/typedef.h
@@ -192,21 +192,4 @@ typedef struct {
 // TODO: move the following to a more appropriate location
 extern int INFO_LEVEL;
 
-#if defined(__GNUC__)
-# define CARAT_CURRENT_FUNCTION __PRETTY_FUNCTION__
-#elif defined(__STDC_VERSION__) && (__STDC_VERSION__ >= 199901)
-#  define CARAT_CURRENT_FUNCTION	__func__
-#elif defined(_MSC_VER)
-#  define CARAT_CURRENT_FUNCTION __FUNCTION__
-#else
-#  define CARAT_CURRENT_FUNCTION "<unknown>"
-#endif
-
-
-extern void *xmalloc_(size_t size, const char *funcname);
-extern void *xrealloc_(void *ptr, size_t size, const char *funcname);
-
-#define xmalloc(size)  xmalloc_(size, CARAT_CURRENT_FUNCTION)
-#define xrealloc(ptr, size)  xrealloc_(ptr, size, CARAT_CURRENT_FUNCTION)
-
 #endif /* CARAT_TYPEDEF_H */
